PR fixes

+++ cpp/include/robotick/framework/common/FixedVector.h
@@ -124,7 +124,8 @@ namespace robotick
 		T& operator[](size_t index)
 		{
 			ROBOTICK_ASSERT_MSG(index < Capacity, "Indexing beyond Capacity (non-const [] accessor)");
-			ROBOTICK_ASSERT_MSG(index < count, "Indexing beyond current size of %u (non-const [] accessor)", count);
+			ROBOTICK_ASSERT_MSG(
+				index < count, "Indexing beyond current size of %zu (non-const [] accessor)", static_cast<size_t>(count));
 			return data_buffer[index];
 		}
 
@@ -134,7 +135,7 @@ namespace robotick
 		const T& operator[](size_t index) const
 		{
 			ROBOTICK_ASSERT_MSG(index < Capacity, "Indexing beyond Capacity (const [] accessor)");
-			ROBOTICK_ASSERT_MSG(index < count, "Indexing beyond current size of %u (const [] accessor)", count);
+			ROBOTICK_ASSERT_MSG(index < count, "Indexing beyond current size of %zu (const [] accessor)", static_cast<size_t>(count));
 			return data_buffer[index];
 		}
